admin-service/requirements.txt
django
djangorestframework
autopep8
pylint
mysql
mysqlclient
pymysql
django-mysql
pillow
django-cors-headers
pika